+++ crates/crb-superagent/src/interplay/subscription.rs
@@ -2,7 +2,7 @@ use super::{Fetcher, Interplay};
 use anyhow::{anyhow, Result};
 use async_trait::async_trait;
 use crb_agent::{Address, Agent, Context, MessageFor};
-use crb_core::UniqueId;
+use crb_core::Unique;
 use crb_send::{Recipient, Sender};
 
 pub trait SubscribeExt<S: Subscription> {
@@ -15,7 +15,7 @@ where
     S: Subscription,
 {
     fn subscribe(&self, subscription: S) -> Fetcher<StateEntry<S>> {
-        let sub_id = UniqueId::new(subscription);
+        let sub_id = Unique::new(subscription);
         let (interplay, fetcher) = Interplay::new_pair(sub_id);
         let msg = Subscribe { interplay };
         let res = self.send(msg);
@@ -41,7 +41,7 @@ pub struct StateEntry<S: Subscription> {
 }
 
 pub struct Entry<S: Subscription> {
-    sub_id: UniqueId<S>,
+    sub_id: Unique<S>,
     recipient: Recipient<Unsubscribe<S>>,
 }
 
@@ -75,17 +75,13 @@ pub trait ManageSubscription<S: Subscription>: Agent {
         msg.interplay.responder.send_result(state_entry)
     }
 
-    async fn subscribe(
-        &mut self,
-        sub_id: UniqueId<S>,
-        _ctx: &mut Context<Self>,
-    ) -> Result<S::State> {
+    async fn subscribe(&mut self, sub_id: Unique<S>, _ctx: &mut Context<Self>) -> Result<S::State> {
         Err(anyhow!(
             "The on_subscribe method in not implemented to handle {sub_id}."
         ))
     }
 
-    async fn unsubscribe(&mut self, sub_id: UniqueId<S>, _ctx: &mut Context<Self>) -> Result<()> {
+    async fn unsubscribe(&mut self, sub_id: Unique<S>, _ctx: &mut Context<Self>) -> Result<()> {
         Err(anyhow!(
             "The on_unsubscribe method in not implemented to handle {sub_id}"
         ))
@@ -93,7 +89,7 @@ pub trait ManageSubscription<S: Subscription>: Agent {
 }
 
 pub struct Subscribe<S: Subscription> {
-    pub interplay: Interplay<UniqueId<S>, StateEntry<S>>,
+    pub interplay: Interplay<Unique<S>, StateEntry<S>>,
 }
 
 #[async_trait]
@@ -108,7 +104,7 @@ where
 }
 
 pub struct Unsubscribe<S: Subscription> {
-    pub sub_id: UniqueId<S>,
+    pub sub_id: Unique<S>,
 }
 
 #[async_trait]